Adding nightly test for upgradability: checks that after proper number of epochs active protocol versoins match
