+++ core/primitives/src/lib.rs
@@ -7,6 +7,7 @@ static ALLOC: jemallocator::Jemalloc = jemallocator::Jemalloc;
 
 pub mod account;
 pub mod block;
+pub mod block_header;
 pub mod challenge;
 pub mod contract;
 pub mod errors;
