+++ nightly/nightly.txt
@@ -32,6 +32,7 @@ pytest --timeout=120 sanity/garbage_collection1.py
 pytest --timeout=300 sanity/gc_after_sync.py
 pytest --timeout=300 sanity/gc_sync_after_sync.py
 pytest --timeout=300 sanity/gc_sync_after_sync.py swap_nodes
+pytest sanity/upgradable.py
 
 # python tests for smart contract deployment and invocation
 pytest contracts/deploy_call_smart_contract.py
