+++ pytest/lib/utils.py
@@ -7,6 +7,7 @@ from rc import run, gcloud
 import os
 import tempfile
 import json
+import time
 from pprint import pprint
 
 
@@ -268,3 +269,17 @@ def collect_gcloud_config(num_nodes):
     with open(outfile, 'w+') as f:
         json.dump(res, f)
     os.environ[CONFIG_ENV_VAR] = outfile
+
+
+def wait_for_blocks_or_timeout(node, num_blocks, timeout, callback=None):
+    status = node.get_status()
+    start_height = status['sync_info']['latest_block_height']
+    max_height = 0
+    started = time.time()
+    while max_height < start_height + num_blocks:
+        assert time.time() - started < timeout
+        status = node.get_status()
+        max_height = status['sync_info']['latest_block_height']
+        if callback is not None:
+            if callback():
+                break
